Dissolve ModeAwareDuckAiChatStore into an ActivityScope resolution of the frozen browser mode
