+++ duckchat/duckchat-impl/src/main/java/com/duckduckgo/duckchat/impl/history/ChatHistoryRepository.kt
@@ -18,7 +18,7 @@ package com.duckduckgo.duckchat.impl.history
 
 import android.content.Context
 import com.duckduckgo.common.utils.DispatcherProvider
-import com.duckduckgo.di.scopes.AppScope
+import com.duckduckgo.di.scopes.ActivityScope
 import com.duckduckgo.duckchat.impl.R
 import com.duckduckgo.duckchat.impl.models.ModelDisplay
 import com.duckduckgo.duckchat.impl.models.toChatType
@@ -50,8 +50,8 @@ interface ChatHistoryRepository {
     suspend fun exportChats(requests: List<ChatExportRequest>): List<File>
 }
 
-@ContributesBinding(AppScope::class)
-@SingleInstanceIn(AppScope::class)
+@ContributesBinding(ActivityScope::class)
+@SingleInstanceIn(ActivityScope::class)
 class RealChatHistoryRepository @Inject constructor(
     private val chatStore: DuckAiChatStore,
     private val dispatchers: DispatcherProvider,
